LRQA-31884 Add more specific exceptions
